refactor: move agent mutations to dedicated module
